// src/types/tmo.ts
export interface TMOAddress {
  id?: number;
  customer_id?: number;
  region?: TMORegion;
  region_id?: number;
  country_id: string;
  street: string[];
  telephone: string;
  postcode: string;
  city: string;
  firstname: string;
  lastname: string;
  default_shipping?: boolean;
  default_billing?: boolean;
  custom_attributes?: Array<{ attribute_code: string; value: any }>;
}

export interface TMORegion {
  region_code: string;
  region: string;
  region_id: number;
}
// ...
export interface TMOGeoItem {
  id: string;
  name: string;
}